feat(sdk-python): add a2ui_prompt() to copilotkit.a2ui module
Move the dynamic A2UI prompt builder into the SDK so it's reusable
across projects. Usage:

    from copilotkit import a2ui
    prompt = a2ui.a2ui_prompt()
    prompt = a2ui.a2ui_prompt(design_guidelines="Custom rules...")

Includes:
- a2ui_prompt() function with DEFAULT_GENERATION_GUIDELINES and
  DEFAULT_DESIGN_GUIDELINES
- a2ui_json_schema.json bundled in the package
- pyproject.toml updated to include JSON files in distribution
- Example updated to import from copilotkit.a2ui instead of local copy
